[New] Ability to add contacts added
The program can now take in user input to create a new contact. New
contacts are added to the end of the display list, which is then sorted,
and can then be selected. Created contacts are currently not saved and
are deleted when the program is closed. Error checking for user inputed
fields is not yet implemented.