[New] Reading and Writing to JSON added
The Jackson-jr library is used to save stored contacts in a JSON file.
On start up the program reads the corresponding JSON if it has
information, and any time contacts are manipulated, the JSON is updated
accordingly.
In addition, Contact's birthday attribute was changed from a Calendar
object to a LocalDate object for easier use.