[New] Error handling added to Home/Cell Number fields
Under the assumption that Home/Cell Number fields are 10 digits, display
error pop-up if the field has an incorrect number of digits. Users
cannot complete the action of adding/editing a contact until the error
is resolved.
Also added the ability to leave these fields blank in cases where the
contact does not have a home/cell number. The corresponding field will
display as 0 in the display panel to signify it's empty.